src/main/java/org/example/smartinventory/service/DepartmentsServiceImpl.java
package org.example.smartinventory.service;

import org.example.smartinventory.entities.DepartmentEntity;
import org.example.smartinventory.repository.DepartmentRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.math.BigDecimal;
import java.time.LocalDateTime;
import java.util.Collection;
import java.util.List;
import java.util.Map;
import java.util.Optional;

@Service
public class DepartmentsServiceImpl implements DepartmentsService
{
    private final DepartmentRepository departmentRepository;

    @Autowired
    public DepartmentsServiceImpl(DepartmentRepository departmentRepository)
    {
        this.departmentRepository = departmentRepository;
    }


    @Override
    public Collection<DepartmentEntity> findAll() {
        return List.of();
    }

    @Override
    public void save(DepartmentEntity departmentEntity) {

    }

    @Override
    public void delete(DepartmentEntity departmentEntity) {

    }

    @Override
    public Optional<DepartmentEntity> findById(Long id) {
        return Optional.empty();
    }

    @Override
    public Collection<DepartmentEntity> findAllById(Iterable<Long> ids) {
        return List.of();
    }

    @Override
    public Optional<DepartmentEntity> findByDepartmentName(String name) {
        return Optional.empty();
    }

    @Override
    public List<DepartmentEntity> findAllActiveDepartments() {
        return List.of();
    }

    @Override
    public List<DepartmentEntity> findByManagerId(int managerId) {
        return List.of();
    }

    @Override
    public List<DepartmentEntity> findDepartmentsCreatedAfter(LocalDateTime date) {
        return List.of();
    }

    @Override
    public List<DepartmentEntity> findDepartmentsUpdatedAfter(LocalDateTime date) {
        return List.of();
    }

    @Override
    public List<DepartmentEntity> searchDepartments(String keyword) {
        return List.of();
    }

    @Override
    public long countEmployeesInDepartment(int departmentId) {
        return 0;
    }

    @Override
    public List<DepartmentEntity> findDepartmentsWithMoreThanXEmployees(int count) {
        return List.of();
    }

    @Override
    public List<DepartmentEntity> findDepartmentsWithNoEmployees() {
        return List.of();
    }

    @Override
    public boolean updateDepartmentStatus(int id, boolean status) {
        return false;
    }

    @Override
    public boolean updateDepartmentManager(int id, int managerId) {
        return false;
    }

    @Override
    public int deleteInactiveDepartments() {
        return 0;
    }

    @Override
    public List<DepartmentEntity> findByAddress(String address) {
        return List.of();
    }

    @Override
    public Map<Boolean, Long> countActiveAndInactiveDepartments() {
        return Map.of();
    }

    @Override
    public List<DepartmentEntity> findDepartmentsSortedByEmployeeCount() {
        return List.of();
    }

    @Override
    public List<DepartmentEntity> findDepartmentsCreatedBetween(LocalDateTime startDate, LocalDateTime endDate) {
        return List.of();
    }

    @Override
    public List<DepartmentEntity> findUpdatedDepartments() {
        return List.of();
    }

    @Override
    public List<DepartmentEntity> findDepartmentsByCustomCriteria(String name, Boolean isActive, Integer managerId) {
        return List.of();
    }

    @Override
    public int bulkUpdateDepartmentAddresses(String oldAddress, String newAddress) {
        return 0;
    }

    @Override
    public DepartmentEntity createDepartment(DepartmentEntity department) {
        return null;
    }

    @Override
    public Optional<DepartmentEntity> updateDepartment(int id, DepartmentEntity updatedDepartment) {
        return Optional.empty();
    }

    @Override
    public boolean deleteDepartment(int id) {
        return false;
    }

    @Override
    public Page<DepartmentEntity> findAllPaginated(Pageable pageable) {
        return null;
    }

    @Override
    public List<DepartmentEntity> findDepartmentsWithEmployeeCountGreaterThan(int count) {
        return List.of();
    }

    @Override
    public double getAverageDepartmentSize() {
        return 0;
    }

    @Override
    public List<DepartmentEntity> findTopNLargestDepartments(int n) {
        return List.of();
    }

    @Override
    public boolean isDepartmentNameUnique(String name) {
        return false;
    }

    @Override
    public List<DepartmentEntity> findDepartmentsWithoutManager() {
        return List.of();
    }

    @Override
    public Map<String, Long> getDepartmentEmployeeDistribution() {
        return Map.of();
    }

    @Override
    public List<DepartmentEntity> findDepartmentsByEmployeeSkill(String skill) {
        return List.of();
    }

    @Override
    public Optional<DepartmentEntity> assignEmployeeToDepartment(int employeeId, int departmentId) {
        return Optional.empty();
    }

    @Override
    public Optional<DepartmentEntity> removeEmployeeFromDepartment(int employeeId, int departmentId) {
        return Optional.empty();
    }

    @Override
    public List<DepartmentEntity> findDepartmentsWithBudgetGreaterThan(BigDecimal budget) {
        return List.of();
    }

    @Override
    public Optional<BigDecimal> getTotalBudgetForAllDepartments() {
        return Optional.empty();
    }

    @Override
    public List<DepartmentEntity> findDepartmentsWithProjectCount(int count) {
        return List.of();
    }
}
